+++ plugins/logs/src/commonTest/kotlin/com/ae/log/logs/LogPluginMigrationTest.kt
@@ -5,7 +5,6 @@ import com.ae.log.AELogTestApi
 import com.ae.log.InternalAELogApi
 import com.ae.log.logs.model.LogSeverity
 import com.ae.log.plugin.Plugin
-import com.ae.log.plugin.PluginContext
 import kotlin.test.AfterTest
 import kotlin.test.Test
 import kotlin.test.assertEquals
@@ -104,22 +103,6 @@ class LogPluginMigrationTest {
         class ImposterPlugin : Plugin {
             override val id = LogPlugin.ID
             override val name = "Imposter"
-
-            override fun onAttach(context: PluginContext) {}
-
-            override fun onStart() {}
-
-            override fun onStop() {}
-
-            override fun onDetach() {}
-
-            override fun onClear() {}
-
-            override fun onOpen() {}
-
-            override fun onClose() {}
-
-            override fun export() = ""
         }
 
         val imposter = ImposterPlugin()