+++ plugins/network/src/commonTest/kotlin/com/ae/log/network/NetworkPluginMigrationTest.kt
@@ -4,7 +4,6 @@ import com.ae.log.AELog
 import com.ae.log.AELogTestApi
 import com.ae.log.InternalAELogApi
 import com.ae.log.plugin.Plugin
-import com.ae.log.plugin.PluginContext
 import kotlin.test.AfterTest
 import kotlin.test.Test
 import kotlin.test.assertEquals
@@ -96,22 +95,6 @@ class NetworkPluginMigrationTest {
         class ImposterPlugin : Plugin {
             override val id = NetworkPlugin.ID
             override val name = "Imposter"
-
-            override fun onAttach(context: PluginContext) {}
-
-            override fun onStart() {}
-
-            override fun onStop() {}
-
-            override fun onDetach() {}
-
-            override fun onClear() {}
-
-            override fun onOpen() {}
-
-            override fun onClose() {}
-
-            override fun export() = ""
         }
 
         AELog.install(ImposterPlugin())